feat: move status to hook
