+++ app/src/components/EthDrive.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { CheckCircle2, File, Folder, Loader2, XCircle } from "lucide-react";
+import { File, Folder } from "lucide-react";
 import Link from "next/link";
 import React, { useCallback, useState } from "react";
 import {
@@ -29,6 +29,7 @@ import { Switch } from "@/components/ui/switch";
 import { useChain } from "@/hooks/useChain";
 import { useDirectory } from "@/hooks/useDirectory";
 import { useDragAndDrop } from "@/hooks/useDragAndDrop";
+import { useTransactionStatus } from "@/hooks/useTransactionStatus";
 import { useWalletConnect } from "@/hooks/useWalletConnect";
 import { dummySignature, request } from "@/lib/alchemy";
 
@@ -72,6 +73,20 @@ export function EthDrive({ path }: { path?: string }) {
     chainConfig: selectedChainConfig,
     chainAddresses: selectedChainAddresses,
   } = useChain(selectedDirectoryChainId);
+
+  const {
+    transactionSteps,
+    accountAbstractionSteps,
+    steps,
+    currentStep,
+    error,
+    transactionHash,
+    setSteps,
+    setCurrentStep,
+    setTransactionHash,
+    setError,
+    getStepIcon,
+  } = useTransactionStatus();
 
   const handleTransactionAsDirectory = useCallback(
     async (callData: Hex) => {
@@ -204,7 +219,7 @@ export function EthDrive({ path }: { path?: string }) {
                 clearInterval(interval);
                 resolve(userOperationReceipt.result.receipt);
               }
-            }, 2000); // Poll every 2 seconds
+            }, 2000);
           });
         };
         const userOperationReceipt = await pollReceipt(requestId);
@@ -334,58 +349,6 @@ export function EthDrive({ path }: { path?: string }) {
   const [createDirectoryName, setCreateDirectoryName] = useState("");
   const [isTransactionStatusModalOpen, setIsTransactionStatusModalOpen] =
     useState(false);
-
-  const transactionSteps = [
-    { key: "checking-network", label: "Checking Network" },
-    {
-      key: "wait-for-user-signature",
-      label: "Waiting for User Signature",
-    },
-    {
-      key: "wait-for-block-confirmation",
-      label: "Waiting for Block Confirmation",
-    },
-    { key: "confirmed", label: "Transaction Confirmed" },
-  ];
-  const accountAbstractionSteps = [
-    { key: "creating-user-operation", label: "Creating User Operation" },
-    {
-      key: "wait-for-user-signature",
-      label: "Waiting for User Signature",
-    },
-    { key: "sending-user-operation", label: "Sending User Operation" },
-    {
-      key: "wait-for-block-confirmation",
-      label: "Waiting for Block Confirmation",
-    },
-    { key: "confirmed", label: "Transaction Confirmed" },
-  ];
-
-  const [steps, setSteps] = useState<{ key: string; label: string }[]>([]);
-  const [currentStep, setCurrentStep] = useState("");
-  const [transactionHash, setTransactionHash] = useState<string>("");
-  const [error, setError] = useState("");
-
-  const getStepIcon = (stepKey: string): JSX.Element => {
-    const currentStepIndex = steps.findIndex(
-      (step) => step.key === currentStep,
-    );
-    const stepIndex = steps.findIndex((step) => step.key === stepKey);
-
-    if (error && stepIndex >= currentStepIndex) {
-      return <XCircle className="w-6 h-6 text-red-500" />;
-    }
-    if (
-      stepIndex < currentStepIndex ||
-      (stepIndex === steps.length - 1 && stepKey === currentStep)
-    ) {
-      return <CheckCircle2 className="w-6 h-6 text-green-500" />;
-    }
-    if (stepKey === currentStep) {
-      return <Loader2 className="w-6 h-6 text-blue-500 animate-spin" />;
-    }
-    return <div className="w-6 h-6 rounded-full border-2 border-gray-300" />;
-  };
 
   return (
     <div className="flex flex-col h-screen">
